hometask-1

+++ hometask-1.js
@@ -1,42 +1,47 @@
 'use strict'
 
 const fs = require('fs'),
-    childProcess = require('child_process');
-    var fileName = process.argv[2],
+    spawn = require('child_process').spawn;
+
+var fileName = process.argv[2],
     func = process.argv[3];
 
-function copyFile() {
-    var now = new Date().toString(),
-        newFileName = fileName.split('.')[0] + '_' + now.split(' ').join('_') + '.' + fileName.split('.')[1];
-
-    console.log('File ' + fileName + ' will be copied to ' + newFileName);
-    return childProcess.spawn('cp', [fileName, newFileName]);
+if (!fileName || !func) {
+    throw('fileName or func is empty');
 }
 
-function deleteFile() {
-    console.log('File ' + fileName + ' will be deleted');
-    return childProcess.spawn('rm', [fileName]);
-}
-
-fs.watch(fileName, function(){
+fs.watch(fileName, function () {
     var action;
 
     console.log(fileName + ' changed!');
 
-    switch(func) {
-        case 'copy':
-            action = copyFile();
-            break;
-        case 'del':
-            action = deleteFile();
-            break;
-        default:
-            console.log(func + ': command not found');
+    if (func === 'copy') {
+        action = copyFile();
+    } else if (func === 'del') {
+        action = deleteFile();
+    } else {
+        console.log(func + ': command not found');
     }
 
     if (action) {
         action.stdout.pipe(process.stdout);
     }
 });
+
+function copyFile() {
+    var now = new Date().toString(),
+        newFileName = fileName.split('.')[0]
+            + '-'
+            + now.split(' ').join('-')
+            + '.' + fileName.split('.')[1];
+
+    console.log('File will be copied to ' + newFileName);
+    return spawn('cp', [fileName, newFileName]);
+}
+
+function deleteFile() {
+    console.log('File will be deleted');
+    return spawn('rm', [fileName]);
+}
 
 console.log('Watching ' + fileName + ' changes');
